requirements.txt
anthropic>=0.45.0
openai>=1.0
pydantic>=2.0
python-dotenv>=1.0
duckduckgo-search>=7.0
yfinance>=0.2.30
requests>=2.31
beautifulsoup4>=4.12
pypdf>=4.0
-e ./model-router
